refactor: separated login and signup page layout according to FSD
