+++ src/widgets/Auth/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginForm } from './ui/LoginForm.vue';
+export { default as SignupForm } from './ui/SignupForm.vue';
